ENGR00303543 usb: ehci-arc: do noop during resume if it is not in host mode
We find a bug that the usb will produce endless interrupt during the
resume after below steps:

- nothing connected at usb otg port, but enable usb gadget driver
- system enters suspend
- plug in usb cable
- system leaves suspend

The reasons of endless interrupt is: the host controller resume routine
leaves the phy low power mode, but device controller resume routine has
still not been called, so the software considers it is still at low power
mode. After this time, the vbus change interrupt comes, the wakeup routine
considers it is an wakeup interrupt, but in fact it is not. So, the endless
interrupt occurs.

The solution of this problem is do noop during host controller resume
routine if current usb mode is not host mode, and the device controller
reusme routine will handle it.

Acked-by: Jason Liu
Signed-off-by: Peter Chen <peter.chen@freescale.com>
(cherry picked from commit 6751d0a0887ecda72d60d391f05f6464d273d2c4)
Signed-off-by: Jon Smith <jtsmith@pdiarm.com>