serial: imx: reset the uart port all the time
Current code resets the uart port only when it supports the irda mode.
In actually, we also need to reset the uart port in the non-irda mode.
A hang was caught in the following case:

UART A transmits data to the other end. But the transmission maybe
terminated. In some corner case, the TX FIFO maybe not empty.

The kernel will hang at the imx_set_termios():
    ............................................................
    while (!(readl(sport->port.membase + USR2) & USR2_TXDC))
	    barrier();
    ............................................................

This patch resets the uart port all the time in the imx_startup().
And fix the hang.

Signed-off-by: Huang Shijie <b32955@freescale.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit c5530168a56384d92ab813f34ba29760c9f93e7f)
Signed-off-by: Jon Smith <jtsmith@pdiarm.com>